requirements.txt
tqdm
opencv-python==4.10.0.84
imageio
imageio[ffmpeg]
plyfile
open3d
ninja
GPUtil
lpips
torchdiffeq
scikit-image==0.19.0
scikit-learn
einops
roma
numpy==1.26.4
